ai-service/requirements.txt
Flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0
numpy==1.24.3
noise==1.2.2